ThetisLink v2.0.3
Multi-tuner release + wire-protocol breaking change.

Wire-protocol `VERSION` u8 bumped from 2 → 3 (S-meter payload
rearranged). v2.0.2 clients against a v2.0.3 server (or vice versa)
receive a localised `ProtocolVersionMismatch` modal ("Server is
too old" / "Client is too old") instead of garbled audio or a silent
connect failure — the server sends a 4-byte back-channel rejection
packet `[MAGIC, VERSION, 0xFF, 0x00]` so the client can detect the
mismatch. Upgrade server, desktop client and Android client together.

Highlights:
- Multi-tuner runtime — up to two physical StockCorner JC-4s / JC-3s
  tuners in parallel via Adafruit MCP2221A USB-HID breakouts (replaces
  the v2.0.2 serial-port RTS/CTS flow). Per-tuner threshold +
  hysteresis sliders on the yellow tune-status wire (1 MΩ + 1 MΩ
  divider, default 2.25 V / 0.50 V). Board scan + serial programming
  UI in the status panel. Automatic USB reconnect with 5 s retry
  interval. Collapsible MCP2221A section with its open/closed state
  persisted across restarts.
- S-meter overhaul — multi-source subscription via
  `rx_channel_sensors_ex` (Sig peak-hold, Avg true-mean, MaxBin),
  S9-frequency band shift via the fork-broadcast `s9_frequency_ex`
  (fallback 50 MHz against stock), TX FWD-power keeps updating under
  Sig/MaxBin selection.
- CTUN coupled-recenter + RX1/RX2 spectrum mirror — both RX spectra
  stay in sync during rapid tuning, Auto FFT retuned to ~25 FPS,
  connect-time balancing so RX1 and RX2 come up at the same moment.
- MIDI client-side VFO coalesce + auto-recenter handshake — extreme
  MIDI wheel input no longer fills Thetis' VFO queue; auto-recenter
  ownership handshake with the Thetis fork now also works when no
  ThetisLink server is connected.
- PA active_pa + drive-snapshot persistence — survives process kill /
  power loss without waiting for the next `start_server()` write.
- Status-panel protocol version-mismatch banner — previously silent,
  now a visible row and ringbuffer entry.
- Cleanup — `assume_tuned` / `TUNER_DONE_ASSUMED` (state 5) / 500 ms
  assume-deadline pad retired. Legacy `tuner_port` / `tuner_enabled`
  COM-port settings UI removed from the Settings tab. Legacy config
  keys silently ignored so existing config files do not break.

Compliance:
- Vendored Rust crate `mcp2221-hal` v0.1.0 (Copyright (c) 2025
  Rob Wells), dual-licensed `MIT OR Apache-2.0`. ThetisLink uses the
  MIT branch (Apache-2.0 is not compatible with `GPL-2.0-only`).
  Attribution in `NOTICE.md`, full MIT text in
  `compliance/THIRD-PARTY-LICENSES.html`, SPDX SBOM in
  `compliance/sbom.spdx.json`.

See `CHANGELOG.md` for the full v2.0.3 entry and earlier history.

+++ sdr-remote-android/android/app/src/main/java/com/sdrremote/viewmodel/SdrViewModel.kt
@@ -79,7 +79,7 @@ class SdrViewModel(application: Application) : AndroidViewModel(application) {
                         frequencyHz = s.frequencyHz.toLong(),
                         frequencyRx2Hz = s.frequencyRx2Hz.toLong(),
                         mode = s.mode.toInt(),
-                        smeter = s.smeter.toInt(),
+                        smeter = s.smeter,
                         powerOn = s.powerOn,
                         txProfile = s.txProfile.toInt(),
                         nrLevel = s.nrLevel.toInt(),
